feat: power recoery rate
